refactor: implement categories management with CRUD operations and enhance category data handling
