+++ mappers/category.mapper.ts
@@ -11,11 +11,18 @@ import {
 } from "../types/models/category.model";
 
 export function mapCategoryDtoToModel(dto: GetCategoryDTO | CreateCategoryResponseDTO): CategoryModel {
+  const rawTotalSpent =
+    (dto as unknown as { totalSpent?: unknown; TotalSpent?: unknown }).totalSpent ??
+    (dto as unknown as { totalSpent?: unknown; TotalSpent?: unknown }).TotalSpent;
+
+  const totalSpent = Number.isFinite(Number(rawTotalSpent)) ? Math.abs(Number(rawTotalSpent)) : 0;
+
   return {
     id: String(dto.id ?? ""),
     label: String(dto.name ?? "Sin categoría").trim() || "Sin categoría",
     icon: String(dto.icon ?? "Tag").trim() || "Tag",
     color: String(dto.color ?? APP_COLORS.actionPrimary).trim() || APP_COLORS.actionPrimary,
+    totalSpent,
   };
 }
